Float

+++ SpatialGDK/Source/SpatialGDK/Public/LoadBalancing/GridBasedLBStrategy.h
@@ -63,8 +63,8 @@ public:
 
 	uint32 GetRows() const { return Rows; }
 	uint32 GetCols() const { return Cols; }
-	uint32 GetWorldWidth() const { return WorldWidth; }
-	uint32 GetWorldHeight() const { return WorldHeight; }
+	float GetWorldWidth() const { return WorldWidth; }
+	float GetWorldHeight() const { return WorldHeight; }
 
 #if WITH_EDITOR
 	virtual void PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent) override;
